// sim/web/src/generated/PadBank.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PadBinding } from "./PadBinding";

/**
 * Face order: cross, circle, square, triangle. D-pad order: up, right, down, left.
 */
export type PadBank = { name: string, face_tap: [PadBinding, PadBinding, PadBinding, PadBinding], face_hold: [PadBinding, PadBinding, PadBinding, PadBinding], dpad: [PadBinding, PadBinding, PadBinding, PadBinding], };